chore: prepare for GLL algorithm
