+++ project/tensor_cfpq.py
@@ -1,85 +1,12 @@
 from typing import Set, Tuple
-from pyformlang.rsa import RecursiveAutomaton
-from pyformlang.cfg import CFG
-from pyformlang.finite_automaton import (
-    NondeterministicFiniteAutomaton,
-    DeterministicFiniteAutomaton,
-)
-from pyformlang.finite_automaton.finite_automaton import to_state, to_symbol
+
 from networkx import DiGraph
 
+from pyformlang.rsa import RecursiveAutomaton
+
+from project.utils import graph_to_nfa, rsm_to_nfa
+
 from project.adjacency_matrix_fa import AdjacencyMatrixFA, intersect_automata, ms_bfs
-from project.utils import graph_to_nfa
-
-
-def cfg_to_rsm(cfg: CFG) -> RecursiveAutomaton:
-    """Transforms the CFG into RSM
-
-    Parameters
-    ----------
-    cfg : :class:`~pyformlang.cfg.CFG`
-        Context-Free Grammar
-
-    Returns
-    -------
-    rsm : :class:`~pyformlang.rsa.RecursiveAutomaton`
-        Recursive State Machine equivalent to the CFG
-    """
-    cfg_str = cfg.to_text()
-    return RecursiveAutomaton.from_text(cfg_str)
-
-
-def ebnf_to_rsm(ebnf: str) -> RecursiveAutomaton:
-    """Transforms the EBNF into RSM
-
-    Parameters
-    ----------
-    ebnf : str
-        Extended Backus-Naur Form
-
-    Returns
-    -------
-    rsm : :class:`~pyformlang.rsa.RecursiveAutomaton`
-        Recursive State Machine equivalent to the EBNF
-    """
-    return RecursiveAutomaton.from_text(ebnf)
-
-
-def rsm_to_nfa(rsm: RecursiveAutomaton) -> NondeterministicFiniteAutomaton:
-    """Transforms the RSM into NFA
-
-    Parameters
-    ----------
-    rsm : :class:`~pyformlang.rsa.RecursiveAutomaton`
-        Recursive State Machine
-
-    Returns
-    -------
-    nfa : :class:`~pyformlang.finite_automaton.NondeterministicFiniteAutomaton`
-        Nondeterministic Finite Automaton equivalent to the RSM
-    """
-    nfa = NondeterministicFiniteAutomaton()
-
-    for var, box in rsm.boxes.items():
-        dfa: DeterministicFiniteAutomaton = box.dfa
-
-        for state in dfa.start_states:
-            state = to_state((var, state))
-            nfa.add_start_state(state)
-
-        for state in dfa.final_states:
-            state = to_state((var, state))
-            nfa.add_final_state(state)
-
-        graph = dfa.to_networkx()
-        for src, dst, sym in graph.edges(data="label"):
-            if sym:
-                src = to_state((var, src))
-                dst = to_state((var, dst))
-                sym = to_symbol(sym)
-                nfa.add_transition(src, sym, dst)
-
-    return nfa
 
 
 def tensor_based_cfpq(
